21.02.25 id 타입을 int에서 integer형으로 변경

+++ src/main/java/com/sbs/untact/controller/UsrArticleController.java
@@ -1,6 +1,5 @@
 package com.sbs.untact.controller;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +10,6 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.sbs.untact.dto.Article;
 import com.sbs.untact.dto.ResultData;
 import com.sbs.untact.service.ArticleService;
-import com.sbs.untact.util.Util;
 
 @Controller
 public class UsrArticleController {
@@ -35,11 +33,11 @@ public class UsrArticleController {
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
 	public ResultData doAdd(String title, String body) {
-	
-		if(title == null) {
+
+		if (title == null) {
 			return new ResultData("F-1", "title을 입력해주세요.");
 		}
-		if(body == null) {
+		if (body == null) {
 			return new ResultData("F-1", "body를 입력해주세요.");
 		}
 
@@ -48,20 +46,37 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(Integer id) {
+		// int 기본타입 -> null이 들어갈 수 없음
+		// Integer 객체타입 -> null이 들어갈 수 있음
+		if (id == null) {
+			return new ResultData("F-1", "id를 입력해주세요.");
+		}
 
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
 			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
 		}
-		
+
 		return articleService.deleteArticle(id);
 	}
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
+	public ResultData doModify(Integer id, String title, String body) {
+		// int 기본타입 -> null이 들어갈 수 없음
+		// Integer 객체타입 -> null이 들어갈 수 있음
+		if (id == null) {
+			return new ResultData("F-1", "id를 입력해주세요.");
+		}
+
+		if (title == null) {
+			return new ResultData("F-1", "title을 입력해주세요.");
+		}
+		if (body == null) {
+			return new ResultData("F-1", "body를 입력해주세요.");
+		}
 
 		Article article = articleService.getArticle(id);
